pinpam: Refactor to modules, envrc
Refactor the pinpam project to be more organized by using modules, and
add .envrc file to automatically load the nix dev environment for code
editors.
