+++ .gitignore
@@ -2,3 +2,4 @@
 pkg/
 ./src/
 pinpam
+.direnv/
