+++ pinpam-core/src/lib.rs
@@ -7,1006 +7,10 @@
 extern crate rust_i18n;
 i18n!("locales", fallback = "en");
 
-use log::{debug, trace, warn};
-use nix::unistd::Uid;
-use std::io::Read;
-use std::{
-    convert::{TryFrom, TryInto},
-    ffi::c_int,
-    fs,
-    path::{Path, PathBuf},
-    str::FromStr,
-};
-use tss_esapi::{
-    abstraction::nv::read_full,
-    attributes::{NvIndexAttributesBuilder, SessionAttributesBuilder},
-    constants::{response_code::Tss2ResponseCodeKind, NvIndexType},
-    handles::{NvIndexHandle, NvIndexTpmHandle, SessionHandle},
-    interface_types::{
-        algorithm::HashingAlgorithm,
-        resource_handles::{NvAuth, Provision},
-        session_handles::PolicySession,
-    },
-    structures::{Auth, MaxNvBuffer, NvPublic},
-    tcti_ldr::DeviceConfig,
-    Context, Error as TssError,
-};
-
-pub type Result<T> = std::result::Result<T, PinError>;
-
-const PIN_NV_INDEX_BASE: u32 = 0x0100_0000;
-const PIN_VERSION_UID_MAX: u32 = 0x007F_FFFF;
-const PIN_VERSION_NV_INDEX_OFFSET: u32 = PIN_VERSION_UID_MAX + 1;
-const PIN_VERSION_CURRENT: u8 = 2;
-const PIN_VERSION_TAG_SIZE: usize = 1;
-const DEFAULT_PINUTIL_PATH: &str = "/usr/bin/pinutil";
-
-#[derive(Debug, serde::Serialize, serde::Deserialize)]
-pub enum PinError {
-    UserNotFound,
-    PermissionDenied,
-    PinAlreadySet,
-    NoPinSet,
-    PinsDontMatch,
-    PinIsLocked,
-    IncorrectPin { locked: bool },
-    PinIsEmpty,
-    PinContainsNonDigits,
-    GetUsernameForUidFailed(u32),
-    CannotDeletePin(DeleteResult),
-    PinTooShort { length: usize, limit: usize },
-    PinTooLong { length: usize, limit: usize },
-    AlreadyProvisioned(u32),
-    UidOverflow(u32),
-    NotProvisioned(u32),
-    TpmError(String),
-    IoError(String),
-    TermIoError(String),
-    PinutilOutputDecodeError(String),
-}
-
-impl From<TssError> for PinError {
-    fn from(err: TssError) -> Self {
-        Self::TpmError(err.to_string())
-    }
-}
-
-impl From<std::io::Error> for PinError {
-    fn from(err: std::io::Error) -> Self {
-        Self::IoError(err.to_string())
-    }
-}
-
-impl From<nix::errno::Errno> for PinError {
-    fn from(err: nix::errno::Errno) -> Self {
-        Self::TermIoError(err.to_string())
-    }
-}
-
-impl std::fmt::Display for PinError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let msg = match self {
-            Self::UserNotFound => t!("user_not_found"),
-            Self::PermissionDenied => t!("permission_denied"),
-            Self::PinAlreadySet => t!("pin_already_set"),
-            Self::NoPinSet => t!("no_pin_set"),
-            Self::PinsDontMatch => t!("pins_dont_match"),
-            Self::PinIsLocked => t!("pin_is_locked"),
-            Self::IncorrectPin { locked: false } => t!("incorrect_pin"),
-            Self::IncorrectPin { locked: true } => t!("incorrect_pin_locked"),
-            Self::PinIsEmpty => t!("pin_is_empty"),
-            Self::PinContainsNonDigits => t!("pin_contains_non_digits"),
-            Self::GetUsernameForUidFailed(uid) => t!("get_username_failed", "uid" => uid),
-            Self::CannotDeletePin(e) => t!("cannot_delete_pin", "error" => e),
-            Self::PinTooShort { length: _, limit } => t!("pin_too_short", "limit" => limit),
-            Self::PinTooLong { length: _, limit } => t!("pin_too_long", "limit" => limit),
-            Self::AlreadyProvisioned(uid) => t!("already_provisioned", "uid" => uid),
-            Self::UidOverflow(uid) => t!("uid_overflow", "uid" => uid),
-            Self::NotProvisioned(uid) => t!("not_provisioned", "uid" => uid),
-            Self::TpmError(tss_error) => t!("tpm_error", "error" => tss_error),
-            Self::IoError(e) => t!("io_error", "error" => e),
-            Self::TermIoError(e) => t!("term_io_error", "error" => e),
-            Self::PinutilOutputDecodeError(e) => t!("pinutil_decode_error", "error" => e),
-        };
-        write!(f, "{msg}")
-    }
-}
-
-#[derive(Debug, Clone, Copy, serde::Serialize, serde::Deserialize)]
-pub struct AttemptInfo {
-    pub used: u32,
-    pub limit: u32,
-}
-
-impl AttemptInfo {
-    pub fn from_pin_data(slot: PinData) -> Self {
-        let used = if slot.pinCount < 0 {
-            0
-        } else {
-            slot.pinCount as u32
-        };
-        let limit = if slot.pinLimit <= 0 {
-            0
-        } else {
-            slot.pinLimit as u32
-        };
-        Self { used, limit }
-    }
-    pub fn locked(&self) -> bool {
-        self.limit > 0 && self.used >= self.limit
-    }
-    pub fn prompt_tuple(&self) -> (u32, u32) {
-        (self.used, self.limit)
-    }
-}
-
-pub fn serialize_error_as_string<T, S>(err: &T, ser: S) -> ::core::result::Result<S::Ok, S::Error>
-where
-    T: std::fmt::Display,
-    S: serde::Serializer,
-{
-    ser.serialize_str(&format!("{err}"))
-}
-
-/// Policy describing acceptable PIN characteristics.
-#[derive(Debug, Clone)]
-pub struct PinPolicy {
-    /// Minimum allowed length (after trimming).
-    pub min_length: usize,
-    /// Optional maximum length.
-    pub max_length: Option<usize>,
-    // Maximum allowed failed attempts before lockout.
-    pub max_attempts: u32,
-    /// Full path to the trusted pinutil binary.
-    pub pinutil_path: PathBuf,
-}
-
-impl Default for PinPolicy {
-    fn default() -> Self {
-        Self {
-            min_length: 4,
-            max_length: Some(8),
-            max_attempts: 3,
-            pinutil_path: PathBuf::from(DEFAULT_PINUTIL_PATH),
-        }
-    }
-}
-
-impl PinPolicy {
-    pub fn new(
-        min_length: usize,
-        max_length: Option<usize>,
-        max_attempts: u32,
-        pinutil_path: PathBuf,
-    ) -> Self {
-        Self {
-            min_length,
-            max_length,
-            max_attempts,
-            pinutil_path,
-        }
-    }
-    pub fn validate(&self, pin: &str) -> Result<()> {
-        let length = pin.len();
-
-        // Verify PIN only contains digits and is not empty after trimming whitespace.
-        let trimmed = pin.trim();
-        if trimmed.is_empty() {
-            return Err(PinError::PinIsEmpty);
-        }
-        if !trimmed.chars().all(|c| c.is_ascii_digit()) {
-            return Err(PinError::PinContainsNonDigits);
-        }
-
-        if length < self.min_length {
-            return Err(PinError::PinTooShort {
-                length,
-                limit: self.min_length,
-            });
-        }
-
-        if let Some(max_len) = self.max_length {
-            if length > max_len {
-                return Err(PinError::PinTooLong {
-                    length,
-                    limit: max_len,
-                });
-            }
-        }
-
-        Ok(())
-    }
-    pub fn parse_config(config: &str) -> Result<Self> {
-        let mut min_length = 4;
-        let mut max_length = Some(8);
-        let mut max_attempts = 3;
-        let mut pinutil_path = PathBuf::from(DEFAULT_PINUTIL_PATH);
-
-        for part in config.split_whitespace() {
-            let mut iter = part.splitn(2, '=');
-            let key = iter.next().unwrap();
-            let value = iter.next().ok_or_else(|| {
-                PinError::from(TssError::WrapperError(
-                    tss_esapi::WrapperErrorKind::ParamsMissing,
-                ))
-            })?;
-
-            match key {
-                "pin_min_length" => {
-                    min_length = value.parse().map_err(|_| {
-                        PinError::from(TssError::WrapperError(
-                            tss_esapi::WrapperErrorKind::InvalidParam,
-                        ))
-                    })?;
-                }
-                "pin_max_length" => {
-                    max_length = Some(value.parse().map_err(|_| {
-                        PinError::from(TssError::WrapperError(
-                            tss_esapi::WrapperErrorKind::InvalidParam,
-                        ))
-                    })?);
-                }
-                "pin_lockout_max_attempts" => {
-                    max_attempts = value.parse().map_err(|_| {
-                        PinError::from(TssError::WrapperError(
-                            tss_esapi::WrapperErrorKind::InvalidParam,
-                        ))
-                    })?;
-                }
-                "pinutil_path" => {
-                    let candidate = PathBuf::from(value);
-                    if !candidate.is_absolute() {
-                        warn!("Ignoring pinutil_path '{}': path must be absolute", value);
-                        return Err(PinError::from(TssError::WrapperError(
-                            tss_esapi::WrapperErrorKind::InvalidParam,
-                        )));
-                    }
-
-                    match fs::metadata(&candidate) {
-                        Ok(metadata) if metadata.is_file() => {
-                            pinutil_path = candidate;
-                        }
-                        Ok(_) => {
-                            warn!("Ignoring pinutil_path '{}': not a regular file", value);
-                            return Err(PinError::from(TssError::WrapperError(
-                                tss_esapi::WrapperErrorKind::InvalidParam,
-                            )));
-                        }
-                        Err(err) => {
-                            warn!(
-                                "Ignoring pinutil_path '{}': metadata lookup failed ({})",
-                                value, err
-                            );
-                            return Err(PinError::from(TssError::WrapperError(
-                                tss_esapi::WrapperErrorKind::InvalidParam,
-                            )));
-                        }
-                    }
-                }
-                _ => {}
-            }
-        }
-
-        Ok(PinPolicy::new(
-            min_length,
-            max_length,
-            max_attempts,
-            pinutil_path,
-        ))
-    }
-
-    /// Load the PIN policy from the standard configuration locations, falling back to defaults.
-    pub fn load_from_standard_locations() -> Self {
-        const PATHS: [&str; 1] = ["/etc/pinpam/policy"];
-        for path in PATHS {
-            if let Some(policy) = Self::load_from_path(path) {
-                return policy;
-            }
-        }
-        PinPolicy::default()
-    }
-
-    /// Attempt to load a PIN policy from a specific path if it passes security checks.
-    pub fn load_from_path<P: AsRef<Path>>(path: P) -> Option<Self> {
-        let path = path.as_ref();
-        let config = read_policy_if_secure(path)?;
-        match PinPolicy::parse_config(&config) {
-            Ok(policy) => Some(policy),
-            Err(err) => {
-                warn!("Failed to parse PIN policy at {}: {}", path.display(), err);
-                None
-            }
-        }
-    }
-}
-
-fn read_policy_if_secure(path: &Path) -> Option<String> {
-    let mut file = fs::File::open(path).ok()?;
-    let metadata = file
-        .metadata()
-        .inspect_err(|err| {
-            warn!(
-                "Failed to read file metadata at {}: {}",
-                path.display(),
-                err
-            )
-        })
-        .ok()?;
-
-    if !metadata.is_file() {
-        warn!(
-            "Ignoring PIN policy at {}: not a regular file",
-            path.display()
-        );
-        return None;
-    }
-
-    if !metadata_is_secure(&metadata, path) {
-        return None;
-    }
-
-    let mut contents = String::new();
-    match file.read_to_string(&mut contents) {
-        Ok(_) => Some(contents),
-        Err(err) => {
-            warn!("Failed to read PIN policy at {}: {}", path.display(), err);
-            None
-        }
-    }
-}
-
-#[cfg(unix)]
-fn metadata_is_secure(metadata: &fs::Metadata, path: &Path) -> bool {
-    use std::os::unix::fs::MetadataExt;
-
-    if metadata.uid() != 0 {
-        warn!(
-            "Ignoring PIN policy at {}: expected owner uid 0 but found {}",
-            path.display(),
-            metadata.uid()
-        );
-        return false;
-    }
-
-    let mode = metadata.mode() & 0o777;
-    if (mode & 0o113) != 0 {
-        warn!(
-            "Ignoring PIN policy at {}: expected permissions <=0644 but found {:03o}",
-            path.display(),
-            mode
-        );
-        return false;
-    }
-
-    true
-}
-
-#[cfg(not(unix))]
-fn metadata_is_secure(_metadata: &fs::Metadata, _path: &Path) -> bool {
-    true
-}
-pub struct PinManager {
-    context: Context,
-    policy: PinPolicy,
-}
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-#[repr(C)]
-#[allow(non_snake_case)]
-pub struct PinData {
-    pub pinCount: c_int,
-    pub pinLimit: c_int,
-}
-
-impl PinData {
-    pub fn new(pin_count: c_int, pin_limit: c_int) -> Self {
-        Self {
-            pinCount: pin_count,
-            pinLimit: pin_limit,
-        }
-    }
-    pub const SIZE: usize = std::mem::size_of::<PinData>();
-}
-
-impl From<PinData> for Vec<u8> {
-    fn from(value: PinData) -> Self {
-        let mut bytes = Vec::with_capacity(PinData::SIZE);
-        bytes.extend_from_slice(&value.pinCount.to_be_bytes());
-        bytes.extend_from_slice(&value.pinLimit.to_be_bytes());
-        bytes
-    }
-}
-
-impl From<&[u8]> for PinData {
-    fn from(bytes: &[u8]) -> Self {
-        let pin_count = c_int::from_be_bytes(bytes[0..4].try_into().unwrap());
-        let pin_limit = c_int::from_be_bytes(bytes[4..8].try_into().unwrap());
-        Self {
-            pinCount: pin_count,
-            pinLimit: pin_limit,
-        }
-    }
-}
-
-/// Result of PIN verification.
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-pub enum VerificationResult {
-    /// PIN verification succeeded. Contains the current PinData with attempt counters.
-    Success(PinData),
-    /// PIN verification failed - incorrect PIN provided. If `locked` is true, the PIN is now
-    /// locked and future attempts with fail with [`VerificationResult::LockedOut`].
-    Invalid { locked: bool },
-    /// User is locked out due to too many failed attempts.
-    LockedOut,
-}
-
-/// Result of authenticated PIN deletion.
-#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-pub enum DeleteResult {
-    /// PIN deletion succeeded.
-    Success,
-    /// PIN deletion failed - incorrect PIN provided. If `locked` is true, the PIN is now locked
-    /// and future attempts with fail with [`DeleteResult::LockedOut`].
-    Invalid { locked: bool },
-    /// User is locked out due to too many failed attempts.
-    LockedOut,
-}
-
-impl std::fmt::Display for DeleteResult {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        let msg = match self {
-            Self::Success => t!("pin_del_success"),
-            Self::Invalid { locked: false } => t!("pin_del_invalid"),
-            Self::Invalid { locked: true } => t!("pin_del_invalid_locked"),
-            Self::LockedOut => t!("pin_del_locked_out"),
-        };
-        write!(f, "{msg}")
-    }
-}
-
-impl PinManager {
-    /// Create a new PinManager with the given policy.
-    pub fn new(policy: PinPolicy) -> Result<Self> {
-        use std::str::FromStr;
-        use tss_esapi::tcti_ldr::TctiNameConf;
-
-        let tcti = TctiNameConf::from_str("device:/dev/tpmrm0")?;
-        let context = Context::new(tcti)?;
-        Ok(Self { context, policy })
-    }
-
-    /// Provision a new PIN for the supplied user, overwriting anything that might exist.
-    pub fn setup_pin(&mut self, uid: u32, pin: String) -> Result<()> {
-        debug!("Setting up PIN for user '{}'.", uid);
-
-        self.policy.validate(&pin)?;
-
-        let nv_index = nv_index_for_uid(uid)?;
-
-        if self.read_pin_slot_owner(nv_index)?.is_some() {
-            return Err(PinError::AlreadyProvisioned(uid));
-        }
-
-        self.define_pin_slot(nv_index, &pin)?;
-        self.restart_context()?;
-        if let Err(err) = self.write_pin_version_tag(uid, PIN_VERSION_CURRENT) {
-            warn!(
-                "Failed to write PIN version tag for user '{}': {}",
-                uid, err
-            );
-            self.restart_context()?;
-            match self.delete_pin_admin(uid) {
-                Ok(()) => {
-                    return Err(PinError::TpmError(format!(
-                        "failed to write PIN version tag: {}",
-                        err
-                    )));
-                }
-                Err(cleanup_err) => {
-                    return Err(PinError::TpmError(format!(
-                        "failed to write PIN version tag ({}) and failed to rollback PIN setup ({})",
-                        err, cleanup_err
-                    )));
-                }
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Delete the stored PIN for a user, requiring PIN authentication.
-    /// This allows a user to delete their own PIN by providing the correct PIN.
-    /// Returns detailed information about the deletion result.
-    /// uses delete_pin_admin internally after verifying the pin
-    pub fn delete_pin_with_auth(&mut self, uid: u32, pin: &str) -> Result<DeleteResult> {
-        match self.verify_pin(uid, pin)? {
-            VerificationResult::Success(_) => {
-                self.delete_pin_admin(uid)?;
-                Ok(DeleteResult::Success)
-            }
-            VerificationResult::Invalid { locked } => Ok(DeleteResult::Invalid { locked }),
-            VerificationResult::LockedOut => Ok(DeleteResult::LockedOut),
-        }
-    }
-
-    /// Delete the stored PIN for a user with administrative privileges (no PIN required).
-    /// This is intended for use by root or system administrators.
-    pub fn delete_pin_admin(&mut self, uid: u32) -> Result<()> {
-        debug!("Administratively deleting PIN for user '{}'.", uid);
-        let nv_index = nv_index_for_uid(uid)?;
-        let _new_nv_index_handle = self
-            .context
-            .tr_from_tpm_public(nv_index.into())
-            .map(NvIndexHandle::from)?;
-        self.execute_with_auth_session(|ctx| {
-            ctx.nv_undefine_space(Provision::Owner, _new_nv_index_handle)
-        })?;
-        self.clear_sessions();
-        if let Err(err) = self.delete_pin_version_tag(uid) {
-            warn!(
-                "Failed to delete PIN version tag for user '{}': {}",
-                uid, err
-            );
-        }
-        Ok(())
-    }
-
-    /// Validate a user-supplied PIN against the sealed TPM value.
-    /// Returns detailed information about the verification result.
-    pub fn verify_pin(&mut self, uid: u32, pin: &str) -> Result<VerificationResult> {
-        trace!("Verifying PIN for user '{}'.", uid);
-        self.policy.validate(pin)?;
-
-        self.clear_sessions();
-        match self.read_pin_version_tag(uid)? {
-            Some(version) => {
-                if version != PIN_VERSION_CURRENT {
-                    return Err(PinError::TpmError(format!(
-                        "unsupported PIN version tag: {}",
-                        version
-                    )));
-                }
-                self.verify_pin_exact(uid, pin)
-            }
-            None => {
-                let legacy_pin = normalize_legacy_pin(pin);
-                let result = self.verify_pin_exact(uid, &legacy_pin)?;
-                if matches!(result, VerificationResult::Success(_)) {
-                    self.restart_context()?;
-                    if let Err(err) = self.migrate_legacy_pin_format(uid, pin) {
-                        warn!(
-                            "Failed to migrate legacy PIN format for user '{}': {}",
-                            uid, err
-                        );
-                    }
-                    self.restart_context()?;
-                }
-                Ok(result)
-            }
-        }
-    }
-
-    fn verify_pin_exact(&mut self, uid: u32, pin: &str) -> Result<VerificationResult> {
-        let nv_index = nv_index_for_uid(uid)?;
-        let _new_nv_index_handle = self
-            .context
-            .tr_from_tpm_public(nv_index.into())
-            .map(NvIndexHandle::from)?;
-
-        self.context
-            .execute_with_nullauth_session(|ctx| {
-                let auth = Auth::try_from(pin.to_string().as_bytes())?;
-                ctx.tr_set_auth(_new_nv_index_handle.into(), auth)?;
-
-                ctx.nv_read(
-                    NvAuth::NvIndex(_new_nv_index_handle),
-                    _new_nv_index_handle,
-                    PinData::SIZE as u16,
-                    0,
-                )
-            })
-            .map(|data| {
-                let slot = PinData::from(data.as_slice());
-                if slot.pinCount >= slot.pinLimit {
-                    VerificationResult::LockedOut
-                } else {
-                    VerificationResult::Success(slot)
-                }
-            })
-            .or_else(|e| match e {
-                TssError::Tss2Error(rc) => match rc.kind() {
-                    Some(Tss2ResponseCodeKind::AuthFail) | Some(Tss2ResponseCodeKind::BadAuth) => {
-                        let locked =
-                            self.read_pin_slot_owner(nv_index)
-                                .map_or(false, |opt_pin_data| {
-                                    opt_pin_data
-                                        .map_or(false, |slot| slot.pinCount >= slot.pinLimit)
-                                });
-                        Ok(VerificationResult::Invalid { locked })
-                    }
-                    Some(Tss2ResponseCodeKind::Handle)
-                    | Some(Tss2ResponseCodeKind::NvUninitialized) => {
-                        Err(PinError::NotProvisioned(uid))
-                    }
-                    _ => Err(PinError::from(TssError::Tss2Error(rc))),
-                },
-                _ => Err(PinError::from(e)),
-            })
-    }
-
-    fn migrate_legacy_pin_format(&mut self, uid: u32, entered_pin: &str) -> Result<()> {
-        let legacy_pin = normalize_legacy_pin(entered_pin);
-
-        if legacy_pin == entered_pin {
-            return self.write_pin_version_tag(uid, PIN_VERSION_CURRENT);
-        }
-
-        self.delete_pin_admin(uid)?;
-        self.restart_context()?;
-        self.setup_pin(uid, entered_pin.to_owned())?;
-        self.restart_context()?;
-        Ok(())
-    }
-
-    fn read_pin_version_tag(&mut self, uid: u32) -> Result<Option<u8>> {
-        let nv_index = version_nv_index_for_uid(uid)?;
-        let result = self
-            .context
-            .execute_with_nullauth_session(|ctx| read_full(ctx, NvAuth::Owner, nv_index));
-        self.clear_sessions();
-
-        match result {
-            Ok(data) => {
-                if data.len() != PIN_VERSION_TAG_SIZE {
-                    return Err(PinError::TpmError(format!(
-                        "invalid PIN version tag payload size: {}",
-                        data.len()
-                    )));
-                }
-                Ok(data.first().copied())
-            }
-            Err(TssError::Tss2Error(rc)) => match rc.kind() {
-                Some(Tss2ResponseCodeKind::Handle)
-                | Some(Tss2ResponseCodeKind::NvUninitialized) => Ok(None),
-                _ => Err(PinError::from(TssError::Tss2Error(rc))),
-            },
-            Err(e) => Err(PinError::from(e)),
-        }
-    }
-
-    fn write_pin_version_tag(&mut self, uid: u32, version: u8) -> Result<()> {
-        let nv_index = version_nv_index_for_uid(uid)?;
-
-        match self.write_pin_version_tag_existing(nv_index, version) {
-            Ok(()) => Ok(()),
-            Err(TssError::Tss2Error(rc)) => match rc.kind() {
-                Some(Tss2ResponseCodeKind::Handle)
-                | Some(Tss2ResponseCodeKind::NvUninitialized) => {
-                    self.define_pin_version_tag(nv_index)?;
-                    self.write_pin_version_tag_existing(nv_index, version)
-                        .map_err(PinError::from)
-                }
-                _ => Err(PinError::from(TssError::Tss2Error(rc))),
-            },
-            Err(e) => Err(PinError::from(e)),
-        }
-    }
-
-    fn write_pin_version_tag_existing(
-        &mut self,
-        nv_index: NvIndexTpmHandle,
-        version: u8,
-    ) -> std::result::Result<(), TssError> {
-        let payload = [version];
-        let result = self.context.execute_with_nullauth_session(|ctx| {
-            let handle = ctx
-                .tr_from_tpm_public(nv_index.into())
-                .map(NvIndexHandle::from)?;
-            ctx.nv_write(
-                NvAuth::Owner,
-                handle,
-                MaxNvBuffer::try_from(payload.as_slice())?,
-                0,
-            )
-        });
-        self.clear_sessions();
-        result
-    }
-
-    fn define_pin_version_tag(&mut self, nv_index: NvIndexTpmHandle) -> Result<()> {
-        self.context.execute_with_nullauth_session(|ctx| {
-            let attributes = NvIndexAttributesBuilder::new()
-                .with_nv_index_type(NvIndexType::Ordinary)
-                .with_owner_read(true)
-                .with_owner_write(true)
-                .with_no_da(true)
-                .build()?;
-            attributes.validate()?;
-
-            let nv_public = NvPublic::builder()
-                .with_nv_index(nv_index)
-                .with_index_name_algorithm(HashingAlgorithm::Sha256)
-                .with_index_attributes(attributes)
-                .with_data_area_size(PIN_VERSION_TAG_SIZE)
-                .build()?;
-
-            ctx.nv_define_space(Provision::Owner, None, nv_public)?;
-            Ok::<(), TssError>(())
-        })?;
-        self.clear_sessions();
-        Ok(())
-    }
-
-    fn delete_pin_version_tag(&mut self, uid: u32) -> Result<()> {
-        let nv_index = version_nv_index_for_uid(uid)?;
-        let result = self.context.execute_with_nullauth_session(|ctx| {
-            let handle = ctx
-                .tr_from_tpm_public(nv_index.into())
-                .map(NvIndexHandle::from)?;
-            ctx.nv_undefine_space(Provision::Owner, handle)
-        });
-        self.clear_sessions();
-
-        match result {
-            Ok(()) => Ok(()),
-            Err(TssError::Tss2Error(rc)) => match rc.kind() {
-                Some(Tss2ResponseCodeKind::Handle)
-                | Some(Tss2ResponseCodeKind::NvUninitialized) => Ok(()),
-                _ => Err(PinError::from(TssError::Tss2Error(rc))),
-            },
-            Err(e) => Err(PinError::from(e)),
-        }
-    }
-    pub fn clear_sessions(&mut self) {
-        self.context.clear_sessions();
-    }
-    pub fn restart_context(&mut self) -> Result<()> {
-        self.context = Context::new(tss_esapi::tcti_ldr::TctiNameConf::Device(
-            DeviceConfig::from_str("/dev/tpmrm0").map_err(|e| PinError::from(e))?,
-        ))
-        .map_err(|e| PinError::from(e))?;
-        Ok(())
-    }
-    /// Report whether a user is currently locked out.
-    pub fn is_locked_out(&mut self, uid: u32) -> Result<bool> {
-        let nv_index = nv_index_for_uid(uid)?;
-        match self.read_pin_slot_owner(nv_index)? {
-            Some(slot) => {
-                if slot.pinCount >= slot.pinLimit {
-                    Ok(true)
-                } else {
-                    Ok(false)
-                }
-            }
-            None => Ok(false),
-        }
-    }
-
-    /// Return the failed-attempt counter for a user.
-    pub fn get_attempt_count(&mut self, uid: u32) -> Result<Option<u32>> {
-        let nv_index = nv_index_for_uid(uid)?;
-        match self.read_pin_slot_owner(nv_index)? {
-            Some(slot) => Ok(Some(slot.pinCount as u32)),
-            None => Ok(None),
-        }
-    }
-
-    /// Return the full PIN slot data for a user if provisioned.
-    pub fn get_pin_slot(&mut self, uid: u32) -> Result<Option<PinData>> {
-        let nv_index = nv_index_for_uid(uid)?;
-        self.read_pin_slot_owner(nv_index)
-    }
-
-    /// Helper method to execute operations within an authenticated HMAC session.
-    /// This ensures proper session management for TPM operations that require authentication.
-    fn execute_with_auth_session<F, T>(&mut self, f: F) -> ::core::result::Result<T, TssError>
-    where
-        F: FnOnce(&mut Context) -> std::result::Result<T, TssError>,
-    {
-        let session = self.context.start_auth_session(
-            None,
-            None,
-            None,
-            tss_esapi::constants::SessionType::Hmac,
-            tss_esapi::structures::SymmetricDefinition::AES_256_CFB,
-            HashingAlgorithm::Sha256,
-        )?;
-        let ret = self.context.execute_with_session(session, f);
-        self.clear_sessions();
-        ret
-    }
-
-    fn define_pin_slot(&mut self, nv_index: NvIndexTpmHandle, pin: &str) -> Result<()> {
-        // Step 2: Apply policy_nv_written to the trial session
-        // This sets up the policy that the NV index must be in the "written" state
-        let auth_value = Auth::try_from(pin.to_string().as_bytes())?;
-        let (nv_public, _) = self.context.execute_without_session(|ctx| {
-            // Step 1: Create a trial policy session to compute the policy digest
-            let trial_session = ctx
-                .start_auth_session(
-                    None,
-                    None,
-                    None,
-                    tss_esapi::constants::SessionType::Trial,
-                    tss_esapi::structures::SymmetricDefinition::AES_256_CFB,
-                    HashingAlgorithm::Sha256,
-                )?
-                .expect("Failed to create trial session");
-            let (policy_auth_session_attributes, policy_auth_session_attributes_mask) =
-                SessionAttributesBuilder::new()
-                    .with_decrypt(true)
-                    .with_encrypt(true)
-                    .build(); //
-            ctx.tr_sess_set_attributes(
-                trial_session,
-                policy_auth_session_attributes,
-                policy_auth_session_attributes_mask,
-            )?;
-
-            let (policy_auth_session_attributes, policy_auth_session_attributes_mask) =
-                SessionAttributesBuilder::new()
-                    .with_decrypt(true)
-                    .with_encrypt(true)
-                    .build(); //
-            let policy_session = PolicySession::try_from(trial_session)?;
-            ctx.tr_sess_set_attributes(
-                tss_esapi::interface_types::session_handles::AuthSession::PolicySession(
-                    policy_session,
-                ),
-                policy_auth_session_attributes,
-                policy_auth_session_attributes_mask,
-            )?;
-            ctx.policy_command_code(policy_session, tss_esapi::constants::CommandCode::NvWrite)?;
-            ctx.policy_nv_written(policy_session, false)?;
-            let digest = ctx.policy_get_digest(policy_session)?;
-            let attributes = NvIndexAttributesBuilder::new()
-                .with_nv_index_type(NvIndexType::PinFail)
-                .with_auth_read(true)
-                .with_owner_read(true)
-                .with_policy_write(true)
-                .with_no_da(true)
-                .build()?;
-            attributes.validate()?;
-            let nv_public = NvPublic::builder()
-                .with_nv_index(nv_index)
-                .with_index_name_algorithm(HashingAlgorithm::Sha256)
-                .with_index_attributes(attributes)
-                .with_data_area_size(PinData::SIZE)
-                .with_index_auth_policy(digest.clone())
-                .build()?;
-            ctx.clear_sessions();
-            ctx.flush_context(SessionHandle::from(trial_session).into())?;
-            Ok::<(NvPublic, tss_esapi::structures::Digest), TssError>((nv_public, digest))
-        })?;
-        self.context.execute_with_nullauth_session(|ctx| {
-            ctx.nv_define_space(Provision::Owner, Some(auth_value.clone()), nv_public)?;
-            Ok::<(), TssError>(())
-        })?;
-
-        self.context.execute_without_session(|ctx| {
-            let auth_session = ctx
-                .start_auth_session(
-                    None,
-                    None,
-                    None,
-                    tss_esapi::constants::SessionType::Policy,
-                    tss_esapi::structures::SymmetricDefinition::AES_256_CFB,
-                    HashingAlgorithm::Sha256,
-                )?
-                .expect("Failed to create auth session");
-            // re-apply the same policy to the auth session
-            let (policy_auth_session_attributes, policy_auth_session_attributes_mask) =
-                SessionAttributesBuilder::new()
-                    .with_decrypt(true)
-                    .with_encrypt(true)
-                    .build(); //
-            ctx.tr_sess_set_attributes(
-                auth_session,
-                policy_auth_session_attributes,
-                policy_auth_session_attributes_mask,
-            )?;
-            let policy_session = PolicySession::try_from(auth_session)?;
-            ctx.tr_sess_set_attributes(
-                tss_esapi::interface_types::session_handles::AuthSession::PolicySession(
-                    policy_session,
-                ),
-                policy_auth_session_attributes,
-                policy_auth_session_attributes_mask,
-            )?;
-            ctx.policy_command_code(policy_session, tss_esapi::constants::CommandCode::NvWrite)?;
-            ctx.policy_nv_written(policy_session, false)?;
-
-            let nv_index_handle = ctx
-                .tr_from_tpm_public(nv_index.into())
-                .map(NvIndexHandle::from)?;
-            ctx.execute_with_session(Some(auth_session), |ctx| {
-                ctx.tr_set_auth(SessionHandle::from(auth_session).into(), auth_value)?;
-                let initial_data = PinData::new(0, self.policy.max_attempts as c_int);
-                let initial_bytes: Vec<u8> = initial_data.into();
-                ctx.nv_write(
-                    NvAuth::NvIndex(nv_index_handle),
-                    nv_index_handle,
-                    MaxNvBuffer::try_from(initial_bytes.as_slice()).unwrap(),
-                    0,
-                )?;
-                Ok(())
-            })?;
-            Ok::<(), TssError>(())
-        })?;
-
-        Ok(())
-    }
-
-    fn read_pin_slot_owner(&mut self, nv_index: NvIndexTpmHandle) -> Result<Option<PinData>> {
-        let result = self.execute_with_auth_session(|ctx| {
-            let data = read_full(ctx, NvAuth::Owner, nv_index)?;
-            let slot = PinData::from(data.as_slice());
-            Ok(slot)
-        });
-
-        // Handle the case where the NV index doesn't exist
-        match result {
-            Ok(slot) => Ok(Some(slot)),
-            Err(TssError::Tss2Error(rc)) => match rc.kind() {
-                Some(Tss2ResponseCodeKind::Handle)
-                | Some(Tss2ResponseCodeKind::NvUninitialized) => Ok(None),
-                _ => Err(PinError::from(TssError::Tss2Error(rc))),
-            },
-            Err(e) => Err(PinError::from(e)),
-        }
-    }
-}
-
-fn nv_index_for_uid(uid: u32) -> Result<NvIndexTpmHandle> {
-    // add uid to base index, checking for collisions
-    let index_value = PIN_NV_INDEX_BASE
-        .checked_add_signed(uid as i32)
-        .ok_or(PinError::UidOverflow(uid))?;
-    NvIndexTpmHandle::new(index_value).map_err(|e| PinError::TpmError(format!("{e}")))
-}
-
-fn version_nv_index_for_uid(uid: u32) -> Result<NvIndexTpmHandle> {
-    if uid > PIN_VERSION_UID_MAX {
-        return Err(PinError::UidOverflow(uid));
-    }
-
-    let index_value = PIN_NV_INDEX_BASE
-        .checked_add(PIN_VERSION_NV_INDEX_OFFSET)
-        .and_then(|base| base.checked_add(uid))
-        .ok_or(PinError::UidOverflow(uid))?;
-    NvIndexTpmHandle::new(index_value).map_err(|e| PinError::TpmError(format!("{e}")))
-}
-
-fn normalize_legacy_pin(pin: &str) -> String {
-    let normalized = pin.trim_start_matches('0');
-    if normalized.is_empty() {
-        "0".to_owned()
-    } else {
-        normalized.to_owned()
-    }
-}
-
-pub fn get_uid() -> u32 {
-    Uid::current().as_raw()
-}
-
-/// Check if the current user can manage the target user's PIN.
-/// Root (uid 0) can manage anyone's PIN, users can only manage their own.
-pub fn can_manage_pin(target_uid: u32) -> bool {
-    let current_uid = get_uid();
-    current_uid == 0 || current_uid == target_uid
-}
-
-/// Get UID from username using nix crate.
-pub fn get_uid_from_username(username: &str) -> Result<u32> {
-    use nix::unistd::User;
-    User::from_name(username)
-        .map_err(|_| PinError::UserNotFound)?
-        .map(|u| u.uid.as_raw())
-        .ok_or(PinError::UserNotFound)
-}
-
-/// Get username from UID using nix crate.
-pub fn get_username_from_uid(uid: u32) -> Option<String> {
-    use nix::unistd::{Uid, User};
-    User::from_uid(Uid::from_raw(uid)).ok()?.map(|u| u.name)
-}
+pub mod pinconstants;
+pub mod pindata;
+pub mod pinerror;
+pub mod pinindex;
+pub mod pinmanager;
+pub mod pinpolicy;
+pub mod util;
